// types/WeatherDataServerResponse.d.ts
export interface WeatherData {
    location: LocationRawClass;
    forecast: WeatherDataForecast;
    profile: Profile;
}

export interface WeatherDataForecast {
    lastUpdated: Date;
    currentCondition: CurrentCondition;
    forecast: ForecastElement[];
    nowcasting: Nowcasting;
    source: Source;
    provider: Provider;
    aqiProvider: AqiProvider;
    alertsInfo: AlertsInfo;
    alertsInfoList: AlertsInfo[];
    contentData: ContentDatum[];
    mapsMetaData: MapsMetaData;
    unit: number;
    commentary: null;
    lifeDaily: LifeDaily;
    stormInfo: null;
}

export interface AlertsInfo {
    id: string;
    title: string;
    event: string;
    eventName: string;
    level: string;
    abbreviation: string[];
    shortCap: string;
    severity: string;
    significance: string;
    credit: string;
    created: Date;
    start: Date;
    end: Date;
    class: string;
    statusText: string;
    safetyGuide: string;
    detailIndex: string;
}

export interface AqiProvider {
    name: string;
}

export interface ContentDatum {
    id: string;
    cid: number;
    ranking: number;
    contenttype: string;
    content: Content[];
}

export interface Content {
    title: string;
    capIcon: string;
    body: string;
    destUrl: string;
    displayType: string;
    renderPlaces: string[];
}

export interface CurrentCondition {
    isAlert: boolean;
    currentTemperature: string;
    currentTemperatureDesc: number;
    pvdrIcon: string;
    symbol: string;
    normalizedSkyCode: string;
    shortCap: string;
    skycode: Precipitation;
    precipitation: Precipitation;
    deepLink: string;
    degreeSetting: string;
    alternateDegreeSetting: string;
    uv: number;
    uvDesc: string;
    feels: number;
    humidity: string;
    windDesc: string;
    windSpeed: string;
    windDir: number;
    dewPoint: string;
    baro: string;
    visiblity: string;
    richCaps: any[];
    aqi: string;
    aqiSeverity: string;
    aqLevel: number;
    primaryPollutant: string;
    isNight: boolean;
    currentRaw: CurrentRaw;
    unitsRaw: UnitsRaw;
    locationRaw: LocationRawClass;
    alertCount: number;
    alertSignificance: string;
}
// ...
export interface LocationRawClass {
    locality: string;
    region: string;
    subRegion: string;
    country: string;
    isoCode: string;
    language: string;
    latitude: string;
    longitude: string;
    displayName: string;
}

export interface Precipitation {
    children: string;
}

export interface UnitsRaw {
    system: string;
    pressure: Pressure;
    temperature: string;
    speed: Speed;
    height: Height;
    distance: string;
    time: string;
}

export enum Height {
    CM = 'cm',
}

export enum Pressure {
    MB = 'mb',
}

export enum Speed {
    KMH = 'km/h',
}

export interface ForecastElement {
    id: string;
    day: Date;
    dayCode: string;
    nightCode: string;
    daySymbol: ConditionSymbol;
    nightSymbol: NightSymbol;
    highTemp: number;
    lowTemp: number;
    highTemperature: string;
    lowTemperature: string;
    image: Image;
    pvdrIcon: string;
    symbol: ConditionSymbol;
    glyphCode: GlyphCode;
    normalizedSkyCode: string;
    precipitation: string;
    nightPrecipitation: string;
    humidity: string;
    humidityLow: string;
    humidityHi: string;
    windSpeed: string;
    windSpeedUnit: Speed;
    airPressure: AirPressure;
    summaries: string[];
    dayNightSummaries: DayNightSummaries;
    almanac: Almanac;
    deepLink: string;
    hourly: Hourly[];
    uv: number;
    uvDesc: UvDesc;
    windMaxRaw: number;
    windMax: string;
    windMaxDir: number;
    cap: Cap;
    nightCap: NightCap;
    dayCap: Cap;
    rainAmount: RainAmount;
    snowAmount: boolean;
}
// ...
export interface LifeDaily {
    days: { [key: string]: Day }[];
    dayhours: any[];
}

export interface Day {
    rankId: number;
    categoryId: string;
    summary: string;
}

export interface MapsMetaData {
    NowcastVectorMap: NowcastVectorMap;
    NowcastVectorMapV3: NowcastVectorMapV3;
    PRateMap: NowcastVectorMap;
    TileImageMetadata: TileImageMetadata;
    AirQualityHeatMap: AirQualityHeatMap;
    SevereWeatherSvg: SevereWeatherSVG;
    PollenCity: Pollen;
    PollenImage: Pollen;
    HurricaneSvg: HurricaneSVG;
    WildFireSvg: NowcastVectorMap;
    SevereWeatherText: NowcastVectorMap;
    TemperatureGrid: AirQualityHeatMap;
    TemperatureGridRR: TemperatureGridRr;
    TemperatureGridCC: TemperatureGridCC;
    AirQualityHeatMapEu: AirQualityHeatMapEu;
    EuPollenCity: EuPollen;
    EuPollenImage: EuPollen;
    ForecastVideo: ForecastVideo;
    ThunderstormPolygonSvg: TemperatureGridRr;
    LightningStrikeSvg: LightningStrikeSVG;
}
// ...
export interface Nowcasting {
    precipitation: number[];
    templateType: string;
    minutesToTransit: number;
    summary: string;
    shortSummary: string;
    taskbarSummary: string;
    horrizonCount: number;
    minutesBetweenHorrizons: number;
    enableRainSignal: boolean;
    raintype: string;
    timestamp: Date;
    weathertype: number;
    nowcastingDistance: number;
    nearbyPrecipitationType: number;
    isRaining: boolean;
    useNowcastSummary: string;
}

export interface Provider {
    name: string;
    url: string;
}

export interface Source {
    id: string;
    coordinates: Coordinates;
    location: SourceLocation;
    utcOffset: string;
    countryCode: string;
}

export interface Coordinates {
    lat: number;
    lon: number;
}

export interface SourceLocation {
    Name: string;
    StateCode: string;
    CountryName: string;
    CountryCode: string;
    TimezoneName: string;
    TimezoneOffset: string;
}

export interface Profile {
    followedLocations: any[];
    detectedLocation: DetectedLocation;
}

export interface DetectedLocation {
    from: number;
    language: string;
    latitude: string;
    longitude: string;
}